fel: Add fel spl command support for Allwinner A64
The SCTLR bits are somewhat different because the V bit is set
to 0 on A64 (Low exception vectors, base address 0x00000000) and
the UNK bit (Reads of this bit return an UNKNOWN value) is also not
the same as on the other SoCs. So the SCTLR check can be relaxed.

Tested using a custom built UART0 hello world bootable image.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>